chore(instrumentation-fs): consolidate common test utilities
