+++ plugins/node/instrumentation-fs/test/utils.ts
@@ -1,0 +1,77 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { ReadableSpan } from '@opentelemetry/sdk-trace-base';
+import * as assert from 'assert';
+import { SpanKind, SpanStatusCode } from '@opentelemetry/api';
+import type { FMember } from '../src/types';
+
+export const assertSpans = (spans: ReadableSpan[], expected: any) => {
+  assert.strictEqual(
+    spans.length,
+    expected.length,
+    `Expected ${expected.length} spans, got ${spans.length}(${spans
+      .map((s: any) => `"${s.name}"`)
+      .join(', ')})`
+  );
+
+  spans.forEach((span, i) => {
+    assertSpan(span, expected[i]);
+  });
+};
+
+const assertSpan = (span: ReadableSpan, expected: any) => {
+  assert(span);
+  assert.strictEqual(span.name, expected.name);
+  assert.strictEqual(
+    span.kind,
+    SpanKind.INTERNAL,
+    'Expected to be of INTERNAL kind'
+  );
+  if (expected.parentSpan) {
+    assert.strictEqual(
+      span.parentSpanId,
+      expected.parentSpan.spanContext().spanId
+    );
+  }
+  if (expected.attributes) {
+    assert.deepEqual(span.attributes, expected.attributes);
+  }
+  if (expected.error) {
+    assert(
+      expected.error.test(span.status.message),
+      `Expected "${span.status.message}" to match ${expected.error}`
+    );
+    assert.strictEqual(span.status.code, SpanStatusCode.ERROR);
+  } else {
+    assert.strictEqual(
+      span.status.code,
+      SpanStatusCode.UNSET,
+      'Expected status to be unset'
+    );
+    assert.strictEqual(span.status.message, undefined);
+  }
+};
+
+export const makeRootSpanName = (name: FMember): string => {
+  let rsn: string;
+  if (Array.isArray(name)) {
+    rsn = `${name[0]}.${name[1]}`;
+  } else {
+    rsn = `${name}`;
+  }
+  rsn = `${rsn} test span`;
+  return rsn;
+};
